Finished up the implementation of the Datapoint class and reworked the project to use arrays instead of vectors

+++ BaosMini/include/Utility/FormatterFunctions.hpp
@@ -7,7 +7,7 @@
 namespace FormatterFunctions
 {
 	bool formatValueInTwoBytes(
-		int value,
+		unsigned short value,
 		unsigned char* firstByteVar,
 		unsigned char* secondByteVar
 	);
