+++ BaosMini/include/telegram/GetServerItem.hpp
@@ -2,13 +2,14 @@
 #define GET_SERVER_ITEM_HPP
 
 #include "telegram/BaosTelegram.hpp"
+#include "Utility/FormatterFunctions.hpp"
 
 class GetServerItem : public BaosTelegram
 {
 public:
-    GetServerItem(
-       int firstItemId,
-       int nrOfItems);
+	GetServerItem(
+		unsigned short firstItemId,
+		unsigned short nrOfItems);
 	~GetServerItem();
 
 private:
